Jy (#44)
* video

* likeList start

* start

* not yet

+++ src/main/java/com/example/MMP/wod/WodController.java
@@ -1,8 +1,12 @@
 package com.example.MMP.wod;
 
+import com.example.MMP.siteuser.SiteUser;
+import com.example.MMP.siteuser.SiteUserService;
 import jakarta.validation.Valid;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.util.StringUtils;
@@ -11,7 +15,9 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.security.Principal;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Controller
 @RequiredArgsConstructor
@@ -19,6 +25,7 @@ import java.util.List;
 public class WodController {
     private final WodService wodService;
     private final FileUploadUtil fileUploadUtil;
+    private final SiteUserService siteUserService;
 
     @GetMapping("/form")
     private String wod(Model model){
@@ -33,12 +40,14 @@ public class WodController {
         return "wod/wod_create";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/create")
-    private String createWod(@Valid WodForm wodForm, BindingResult bindingResult, @RequestParam("image") MultipartFile image, Model model){
+    private String createWod(@Valid WodForm wodForm, BindingResult bindingResult, @RequestParam("image") MultipartFile image, Model model, Principal principal){
         if (bindingResult.hasErrors()) {
             return "wod/wod_create";
         }
 
+        SiteUser writer = siteUserService.getUser(principal.getName());
         // 이미지 파일의 경로가 비어있지 않으면 업로드를 시도합니다.
         if (image != null && !image.isEmpty()) {
             String fileName = StringUtils.cleanPath(image.getOriginalFilename());
@@ -54,7 +63,7 @@ public class WodController {
 
                 return "wod/wod_create"; // 업로드 실패 시 처리하는 방법에 따라 변경
             }
-            wodService.create(fileName, wodForm.getContent());
+            wodService.create(fileName, wodForm.getContent(), writer);
         }
 
         List<Wod> wodList = wodService.getList();
@@ -71,9 +80,13 @@ public class WodController {
         return "wod/wod_detail";
     }
 
-    @PostMapping("/delete/{id}")
-    public String delete(@PathVariable("id") Long id){
-        wodService.delete(id);
+    @PreAuthorize("isAuthenticated()")
+    @GetMapping("/delete/{id}")
+    public String delete(@PathVariable("id") Long id, Principal principal){
+        Wod wod = wodService.getWod(id);
+        if (wod.getWriter().getUserId().equals(principal.getName())){
+            wodService.delete(id);
+        }
         return "redirect:/wod/form";
     }
 
@@ -86,12 +99,37 @@ public class WodController {
         return "wod/wod_create";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/update/{id}")
-    public String update(@Valid WodForm wodForm, BindingResult bindingResult, @PathVariable("id") Long id){
+    public String update(@Valid WodForm wodForm, BindingResult bindingResult, @PathVariable("id") Long id, Principal principal){
         if (bindingResult.hasErrors()) {
             return "wod/wod_create";
         }
-        wodService.update(id, wodForm.getContent());
+        Wod wod = wodService.getWod(id);
+        if (wod.getWriter().getUserId().equals(principal.getName())){
+            wodService.update(id, wodForm.getContent());
+        }
         return "redirect:/wod/detail/" + id;
+    }
+
+
+    @PostMapping("/like")
+    public ResponseEntity<Map<String, Object>> likeWod(@RequestBody Map<String, Long> payload, Principal principal) {
+        Long wodId = payload.get("wodId");
+        int likeListSize = wodService.addLike(wodId, principal.getName());
+        Map<String, Object> response = new HashMap<>();
+        response.put("success", true);
+        response.put("likeListSize", likeListSize);
+        return ResponseEntity.ok(response);
+    }
+
+    @DeleteMapping("/like")
+    public ResponseEntity<Map<String, Object>> unlikeWod(@RequestBody Map<String, Long> payload, Principal principal) {
+        Long wodId = payload.get("wodId");
+        int likeListSize = wodService.removeLike(wodId, principal.getName());
+        Map<String, Object> response = new HashMap<>();
+        response.put("success", true);
+        response.put("likeListSize", likeListSize);
+        return ResponseEntity.ok(response);
     }
 }
